+++ docs/changelog.rst
@@ -6,9 +6,11 @@ Here you can see the list of changes between each Kyoukai release.
 Version 1.9.2 (Unreleased)
 --------------------------
 
- - Add depth property which signifies how deep in the tree the Blueprint is.
+ - Add ``depth`` property which signifies how deep in the tree the Blueprint is.
 
  - The routing tree no longer considers matching routes that don't start with the prefix of the blueprint.
+
+ - Add ``tree_path`` property which shows the full tree path to a Blueprint.
 
 Version 1.9.1
 -------------